requirements.txt
asgiref==3.9.2
Django==5.2.6
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
requests
